Remove commented code, cleanup unused code, move cloud provider constant
