+++ deepfence_server/handler/cloud_node.go
@@ -1,30 +1,16 @@
 package handler
 
 import (
-	"encoding/json"
 	"fmt"
-	"github.com/deepfence/ThreatMapper/deepfence_server/reporters"
-	"github.com/deepfence/ThreatMapper/deepfence_utils/directory"
-	"github.com/sirupsen/logrus"
 	"net/http"
-	"time"
 
 	"github.com/deepfence/ThreatMapper/deepfence_server/model"
-	ctl "github.com/deepfence/ThreatMapper/deepfence_utils/controls"
+	"github.com/deepfence/ThreatMapper/deepfence_server/reporters"
+	"github.com/deepfence/ThreatMapper/deepfence_utils/directory"
 	"github.com/deepfence/ThreatMapper/deepfence_utils/log"
 	"github.com/deepfence/ThreatMapper/deepfence_utils/utils"
 	httpext "github.com/go-playground/pkg/v5/net/http"
-)
-
-//func scanId(req model.ScanTriggerReq) string {
-//	return fmt.Sprintf("%s-%d", req.NodeId, time.Now().Unix())
-//}
-
-const (
-	CloudComplianceRefreshInventory  = "CLOUD_COMPLIANCE_REFRESH_INVENTORY"
-	CloudComplianceScanNodesCacheKey = "CLOUD_COMPLIANCE_NODES_LIST"
-	// FilterTypeCloudtrailTrail        = "cloudtrail_trail"
-	// PendingCloudComplianceScansKey = "PENDING_CLOUD_COMPLIANCE_SCANS_KEY"
+	"github.com/sirupsen/logrus"
 )
 
 func (h *Handler) RegisterCloudNodeAccountHandler(w http.ResponseWriter, r *http.Request) {
@@ -37,33 +23,12 @@ func (h *Handler) RegisterCloudNodeAccountHandler(w http.ResponseWriter, r *http
 
 	monitoredAccountIds := req.MonitoredAccountIds
 	orgAccountId := req.OrgAccountId
-	updatedAtTimestamp := time.Now().Unix()
 	scanList := make(map[string]model.CloudComplianceScanDetails)
 	cloudtrailTrails := make([]model.CloudNodeCloudtrailTrail, 10)
 	nodeId := req.NodeId
 
 	ctx := directory.NewContextWithNameSpace(directory.NonSaaSDirKey)
-	redisClient, err := directory.RedisClient(ctx)
-	if err != nil {
-		return
-	}
 
-	logrus.Debugf("Getting CLOUD_COMPLIANCE_REFRESH_INVENTORY value from Redis for node: %s", nodeId)
-	//doRefresh, err := redisClient.HGet(ctx, CloudComplianceRefreshInventory, nodeId).Result()
-	//if err != nil {
-	//	return
-	//}
-
-	//if doRefresh == "" {
-	//	doRefresh = "false"
-	//}
-	//if doRefresh == "true" {
-	//	logrus.Debugf("Setting CLOUD_COMPLIANCE_REFRESH_INVENTORY value from Redis for node: %s", nodeId)
-	//	err := redisClient.HSet(ctx, CloudComplianceRefreshInventory, nodeId, "false").Err()
-	//	if err != nil {
-	//		return
-	//	}
-	//}
 	doRefresh := "false"
 
 	logrus.Debugf("Monitored account ids count: %d", len(monitoredAccountIds))
@@ -83,122 +48,17 @@ func (h *Handler) RegisterCloudNodeAccountHandler(w http.ResponseWriter, r *http
 		if err != nil {
 			complianceError(w, err.Error())
 		}
-		//jsonNode, err := json.Marshal(node)
-		//if err != nil {
-		//	complianceError(w, err.Error())
-		//}
-		//redisClient.HSet(ctx, CloudComplianceScanNodesCacheKey, node["node_id"], string(jsonNode))
-		//pgClient, err := directory.PostgresClient(ctx)
-		//if err != nil {
-		//	httpext.JSON(w, http.StatusInternalServerError, model.Response{Success: false, Message: err.Error()})
-		//	return
-		//}
-		//var cloudtrailAlertsNotifications []postgresql_db.CloudtrailAlertNotification
-		//var accountTrailsMap map[string][]string
-		//if req.CloudProvider == "aws" {
-		//	cloudtrailAlertsNotifications, err = pgClient.GetCloudtrailAlertNotifications(ctx)
-		//	if err != nil && !errors.Is(err, sql.ErrNoRows) {
-		//		complianceError(w, err.Error())
-		//		return
-		//	}
-		//}
 		for monitoredAccountId, monitoredNodeId := range monitoredAccountIds {
 			var monitoredNode map[string]interface{}
-			//complianceScanNodeDetailsStr, err := redisClient.HGet(ctx, CloudComplianceScanNodesCacheKey, monitoredNodeId).Result()
-			//if err != nil {
-			//	complianceError(w, err.Error())
-			//}
-			//if complianceScanNodeDetailsStr != "" {
-			//	err = json.Unmarshal([]byte(complianceScanNodeDetailsStr), &monitoredNode)
-			//}
-			//monitoredNodeUpdatedAt, ok := monitoredNode["updated_at"].(int64)
-			//if !ok {
-			//	monitoredNodeUpdatedAt = updatedAtTimestamp
-			//}
-			//if len(monitoredNode) > 0 && updatedAtTimestamp > monitoredNodeUpdatedAt {
-			//	monitoredNode["updated_at"] = updatedAtTimestamp
-			//} else if len(monitoredNode) == 0 {
 			monitoredNode = map[string]interface{}{
 				"node_id":        monitoredNodeId,
 				"cloud_provider": req.CloudProvider,
 				"node_name":      monitoredAccountId,
 			}
-			//}
-			//jsonMonitoredNode, err := json.Marshal(node)
-			//if err != nil {
-			//	complianceError(w, err.Error())
-			//}
-			//redisClient.HSet(ctx, CloudComplianceScanNodesCacheKey, monitoredNodeId, string(jsonMonitoredNode))
 			err = model.UpsertCloudComplianceNode(ctx, monitoredNode)
 			if err != nil {
 				complianceError(w, err.Error())
 			}
-			//c := model.CloudNodeAccount{NodeID: monitoredNodeId}
-			//err = c.LoadFromDbByNodeId(ctx, pgClient)
-			//if errors.Is(err, sql.ErrNoRows) {
-			//	c = model.CloudNodeAccount{
-			//		NodeID:        monitoredNodeId,
-			//		NodeName:      monitoredAccountId,
-			//		CloudProvider: req.CloudProvider,
-			//		OrgAccountID:  sql.NullString{String: fmt.Sprintf("aws-%s;<cloud_org>", orgAccountId), Valid: true},
-			//	}
-			//	_, err := c.Create(ctx, pgClient)
-			//	if err != nil {
-			//		httpext.JSON(w, http.StatusInternalServerError, model.Response{Success: false, Message: err.Error()})
-			//		return
-			//	}
-			//} else if err != nil {
-			//	httpext.JSON(w, http.StatusInternalServerError, model.Response{Success: false, Message: err.Error()})
-			//	return
-			//}
-			//if req.CloudProvider == "aws" {
-			//	for _, notification := range cloudtrailAlertsNotifications {
-			//		var notificationFilters map[string]interface{}
-			//		notificationFiltersByte, err := notification.Filters.RawMessage.MarshalJSON()
-			//		if err != nil {
-			//			httpext.JSON(w, http.StatusInternalServerError, model.Response{Success: false, Message: err.Error()})
-			//			return
-			//		}
-			//		err = json.Unmarshal(notificationFiltersByte, &notificationFilters)
-			//		if err != nil {
-			//			httpext.JSON(w, http.StatusInternalServerError, model.Response{Success: false, Message: err.Error()})
-			//			return
-			//		}
-			//		trailFilters, exists := notificationFilters[FilterTypeCloudtrailTrail]
-			//		if !exists {
-			//			continue
-			//		}
-			//		trail, ok := trailFilters.(string)
-			//		if !ok {
-			//			trail = ""
-			//		}
-			//		accountAndTrail := strings.SplitN(trail, "/", 2)
-			//		accountId := accountAndTrail[0]
-			//		trailName := accountAndTrail[1]
-			//		if accountId != monitoredAccountId {
-			//			continue
-			//		}
-			//		trailNodeId := fmt.Sprintf("aws-%s;<cloud_account>", accountId)
-			//		accountTrailNames, exists := accountTrailsMap[trailNodeId]
-			//		if !exists {
-			//			accountTrailNames = make([]string, len(cloudtrailAlertsNotifications))
-			//		}
-			//		trailNameFound := false
-			//		for _, accountTrailName := range accountTrailNames {
-			//			if accountTrailName == trailName {
-			//				trailNameFound = true
-			//			}
-			//		}
-			//		if !trailNameFound {
-			//			trailItem := model.CloudNodeCloudtrailTrail{
-			//				AccountId: accountId,
-			//				TrailName: trailName,
-			//			}
-			//			cloudtrailTrails = append(cloudtrailTrails, trailItem)
-			//			accountTrailNames = append(accountTrailNames, trailName)
-			//		}
-			//	}
-			//}
 			pendingScansList, err := reporters.GetPendingScansList(ctx, utils.CLOUD_COMPLIANCE_SCAN, monitoredNodeId)
 			if err != nil {
 				continue
@@ -211,44 +71,10 @@ func (h *Handler) RegisterCloudNodeAccountHandler(w http.ResponseWriter, r *http
 				}
 				scanList[scan.ScanId] = scanDetail
 			}
-			//currentPendingScansStr, err := redisClient.HGet(ctx, PendingCloudComplianceScansKey, monitoredNodeId).Result()
-			//if err != nil {
-			//	continue
-			//}
-			//if currentPendingScansStr == "" {
-			//	continue
-			//}
-			//var currentPendingScans []model.PendingCloudComplianceScan
-			//err = json.Unmarshal([]byte(currentPendingScansStr), &currentPendingScans)
-			//if err != nil {
-			//	continue
-			//}
-			//pendingScansAvailable := false
-			//for _, scan := range currentPendingScans {
-			//	#     filters = {
-			//		#         "node_id": monitored_node_id,
-			//		#         "scan_id": scan["scan_id"],
-			//		#         "scan_status": ["IN_PROGRESS", "ERROR", "COMPLETED"]
-			//	#     }
-			//
-			//}
-			//#     filters = {
-			//	#         "node_id": monitored_node_id,
-			//	#         "scan_id": scan["scan_id"],
-			//	#         "scan_status": ["IN_PROGRESS", "ERROR", "COMPLETED"]
-			//#     }
-			//#     compliance_log = ESConn.search_by_and_clause(CLOUD_COMPLIANCE_LOGS_INDEX, filters, size=1)
-			//#     if not compliance_log.get("hits", []):
-			//#         pending_scans_available = True
-			//#         scan_list[scan["scan_id"]] = scan
-			//# if not pending_scans_available:
-			//#     redis.hset(PENDING_CLOUD_COMPLIANCE_SCANS_KEY, monitored_node_id, "")
-
 		}
 	} else {
 		logrus.Debugf("Single account monitoring for node: %s", nodeId)
 		node := map[string]interface{}{
-			//"node_id":        fmt.Sprintf("%s-%s-cloud-acc", req.CloudProvider, req.CloudAccount),
 			"node_id":        nodeId,
 			"cloud_provider": req.CloudProvider,
 			"node_name":      req.CloudAccount,
@@ -282,35 +108,6 @@ func (h *Handler) RegisterCloudNodeAccountHandler(w http.ResponseWriter, r *http
 		model.CloudNodeAccountRegisterResp{Data: model.CloudNodeAccountRegisterRespData{Scans: scanList,
 			CloudtrailTrails: cloudtrailTrails, Refresh: doRefresh}})
 	return
-
-	//scanId := scanId(req)
-
-	//binArgs := map[string]string{
-	//	"scan_id":   scanId,
-	//	"hostname":  req.NodeId,
-	//	"node_type": req.ResourceType,
-	//	"node_id":   req.ResourceId,
-	//}
-
-	//internal_req := ctl.StartSecretScanRequest{
-	//	ResourceId:   req.ResourceId,
-	//	ResourceType: ctl.StringToResourceType(req.ResourceType),
-	//	BinArgs:      binArgs,
-	//	Hostname:     req.NodeId,
-	//}
-
-	//b, err := json.Marshal(internal_req)
-	//if err != nil {
-	//	httpext.JSON(w, http.StatusInternalServerError, model.Response{Success: false})
-	//	return
-	//}
-
-	//action := ctl.Action{
-	//	ID:             ctl.StartVulnerabilityScan,
-	//	RequestPayload: string(b),
-	//}
-	//
-	//startScan(w, r, utils.NEO4J_VULNERABILITY_SCAN, scanId, req.NodeId, action)
 }
 
 func (h *Handler) ListCloudNodeAccountHandler(w http.ResponseWriter, r *http.Request) {
@@ -323,7 +120,7 @@ func (h *Handler) ListCloudNodeAccountHandler(w http.ResponseWriter, r *http.Req
 		return
 	}
 
-	if ctl.StringToCloudProvider(req.CloudProvider) == -1 {
+	if utils.StringToCloudProvider(req.CloudProvider) == -1 {
 		err = fmt.Errorf("unknown CloudProvider: %s", req.CloudProvider)
 		log.Error().Msgf("%v", err)
 		httpext.JSON(w, http.StatusBadRequest, model.Response{Success: false, Data: err.Error()})
@@ -360,7 +157,7 @@ func extractCloudNodeDetails(w http.ResponseWriter, r *http.Request) (model.Clou
 		return req, err
 	}
 
-	if ctl.StringToCloudProvider(req.CloudProvider) == -1 {
+	if utils.StringToCloudProvider(req.CloudProvider) == -1 {
 		err = fmt.Errorf("unknown CloudProvider: %s", req.CloudProvider)
 		log.Error().Msgf("%v", err)
 		httpext.JSON(w, http.StatusBadRequest, model.Response{Success: false, Data: err.Error()})
